backend/pyproject.toml
[project]
name = "forecast-studio-backend"
version = "0.1.0"
description = "Forecast Studio — FastAPI backend"
requires-python = ">=3.12,<3.13"
dependencies = [
    "forecaster",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "azure-storage-blob>=12.0.0",
    "azure-identity>=1.0.0",
    "httpx>=0.28.0",
    "joblib>=1.4.0",
    "langfuse>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.11.0",
]

[tool.uv.sources]
forecaster = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
# Imports must follow sys.path bootstrap.
"alembic/env.py" = ["E402"]
"scripts/seed_admin.py" = ["E402"]
